requirements.txt
# 個人情報等重要情報の分離
python-decouple
# 計算ツール
numpy
pandas
# データ取得
oandapyV20
# database
sqlalchemy
# Server

